# perf: improve getTemplate() when cache is enabled