--- a/benchmark/index.ts
+++ b/benchmark/index.ts
@@ -1,11 +1,13 @@
 import output from './output'
 import tag from './tag'
 import demo from './demo'
+import layout from './layout'
 
 async function main () {
   await output()
   await tag()
   await demo()
+  await layout()
 }
 
 main()