--- a/benchmark/templates/layout.liquid
+++ b/benchmark/templates/layout.liquid
@@ -1,0 +1,2 @@
+This is a barely empty layout with a body:
+{% block body %}{% endblock %}