Checking for file existence is unnecessary (and consumes IO) when cache already contains entry.

Contributed a little benchmark suite that measures the improvement:

Before:
```
--- layout ---
cache=false x 2,013 ops/sec ±12.06% (55 runs sampled)
cache=true x 4,865 ops/sec ±8.44% (67 runs sampled)
```
After:
```
--- layout ---
cache=false x 2,149 ops/sec ±11.38% (57 runs sampled)
cache=true x 7,716 ops/sec ±2.93% (70 runs sampled)
```